Solve timezone-related day issue
Solved an issue where the start or end time of an activity could be imprecise by one day, one month or even one year. This is not complete as I still need to implement the timezone selection by using the name of the zone and not the number of seconds of offset.